requirement_libraries.txt
praw
gspread
google-api-python-client
oauth2client
python-dotenv
pandas
matplotlib
seaborn
requests